backend/requirements.txt
Flask
Flask-SQLAlchemy
mysql-connector-python
flask-cors
python-dotenv
gunicorn
Flask-Session